more registration and prep
