+++ packages/server/lib/controllers/connection.controller.ts
@@ -3,7 +3,7 @@ import { envs, logContextGetter } from '@nangohq/logs';
 import { NangoError, accountService, configService, connectionService, errorManager, getProvider, githubAppClient } from '@nangohq/shared';
 import { flags } from '@nangohq/utils';
 
-import { preConnectionDeletion } from '../hooks/connection/on/connection-deleted.js';
+import { preConnectionDeletion } from '../hooks/connection/on/pre-connection-deletion.js';
 import {
     connectionCreated as connectionCreatedHook,
     connectionCreationStartCapCheck as connectionCreationStartCapCheckHook,
