+++ packages/server/lib/hooks/connection/on/post-connection-creation.ts
@@ -1,0 +1,63 @@
+import { OtlpSpan, defaultOperationExpiration } from '@nangohq/logs';
+import { onEventScriptService } from '@nangohq/shared';
+
+import { getOrchestrator } from '../../../utils/utils.js';
+
+import type { LogContextGetter } from '@nangohq/logs';
+import type { RecentlyCreatedConnection } from '@nangohq/types';
+
+export async function postConnectionCreation(
+    createdConnection: RecentlyCreatedConnection,
+    provider: string,
+    logContextGetter: LogContextGetter
+): Promise<void> {
+    if (!createdConnection) {
+        return;
+    }
+    const { environment, account, connection } = createdConnection;
+    const { config_id } = connection;
+
+    if (!config_id || !connection.id) {
+        return;
+    }
+
+    const event = 'post-connection-creation';
+
+    const postConnectionCreationScripts = await onEventScriptService.getByConfig(config_id, event);
+
+    if (postConnectionCreationScripts.length === 0) {
+        return;
+    }
+
+    for (const script of postConnectionCreationScripts) {
+        const { name, file_location: fileLocation, version } = script;
+
+        const logCtx = await logContextGetter.create(
+            { operation: { type: 'events', action: 'post_connection_creation' }, expiresAt: defaultOperationExpiration.action() },
+            {
+                account,
+                environment,
+                integration: { id: config_id, name: connection.provider_config_key, provider: provider },
+                connection: { id: connection.id, name: connection.connection_id },
+                syncConfig: { id: script.id, name: script.name },
+                meta: { event }
+            }
+        );
+        logCtx.attachSpan(new OtlpSpan(logCtx.operation));
+
+        const res = await getOrchestrator().triggerOnEventScript({
+            accountId: account.id,
+            connection: createdConnection.connection,
+            version,
+            name,
+            fileLocation,
+            sdkVersion: script.sdk_version,
+            async: true,
+            logCtx
+        });
+
+        if (res.isErr()) {
+            await logCtx.failed();
+        }
+    }
+}
